Added verification when loading new shaders

+++ Engine/KeyCodes.cpp
@@ -144,7 +144,7 @@ namespace Engine
 
 
 
-    using MouseCode = uint16_t;
+    using MouseCode = uint8_t;
 
     namespace Mouse
     {
